refactor(client/assets): use alias import for client assets
