+++ client/src/components/RichText/RichTextEditor.styles.scss
@@ -1,9 +1,7 @@
-
-
 .rdw-editor-wrapper {
   .rdw-link-decorator-wrapper {
     a {
-      span[data-text=true] {
+      span[data-text='true'] {
         text-decoration: underline !important;
       }
     }
@@ -17,7 +15,7 @@
         background-size: 12px;
         width: 12px;
         height: 12px;
-        content: "";
+        content: '';
       }
     }
     &:focus {
@@ -27,14 +25,15 @@
     &:hover {
       color: var(--chakra-colors-primary-600);
       > ::after {
-        background: transparent url(../../assets/link-hint-hover.svg) 0 0 no-repeat;
+        background: transparent url(../../assets/link-hint-hover.svg) 0 0
+          no-repeat;
         text-decoration: none !important;
         display: inline-block;
         margin-left: 4px;
         background-size: 12px;
         width: 12px;
         height: 12px;
-        content: "";
+        content: '';
       }
     }
   }
